add required positional argument: 'features_dtype' (#5158)

+++ tensorflow/examples/tutorials/monitors/iris_monitors.py
@@ -29,9 +29,9 @@ IRIS_TEST = "iris_test.csv"
 
 # Load datasets.
 training_set = tf.contrib.learn.datasets.base.load_csv_with_header(
-    filename=IRIS_TRAINING, target_dtype=np.int)
+    filename=IRIS_TRAINING, target_dtype=np.int, features_dtype=np.float)
 test_set = tf.contrib.learn.datasets.base.load_csv_with_header(
-    filename=IRIS_TEST, target_dtype=np.int)
+    filename=IRIS_TEST, target_dtype=np.int, features_dtype=np.float)
 
 validation_metrics = {"accuracy": tf.contrib.metrics.streaming_accuracy,
                       "precision": tf.contrib.metrics.streaming_precision,
